% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePrimerPickingFile.R
\name{MakePrimerPickingFile}
\alias{MakePrimerPickingFile}
\title{MakePrimerPickingFile}
\usage{
MakePrimerPickingFile(
  library_layout,
  primer_set = "neb",
  primer_volume = 5/3,
  output_path = ".",
  output_file
)
}
\arguments{
\item{library_layout}{A \code{data.table} with columns \code{library_id},
\code{i7_primer}, \code{"i5_primer"} and EITHER (\code{lib_row} AND
\code{lib_col}) OR \code{lib_well}. If there are multiple plates of library
preps, specify them in the \code{lib_plate} column. A separate file will be
written for each plate.}

\item{primer_set}{Which set of barcodes to use. Only supports NEBNext
Multiplex Oligos for Illumina so far (\code{neb}).}

\item{output_path}{Folder to write the output files}

\item{output_file}{Prefix for the output files. Output files will be named
\code{output_file}.csv, or if there are multiple plates,
\code{output_file}_{plate_name}.csv.}
}
\value{
Writes files to the \code{output_path}
}
\description{
Make an epMotion-compatible csv for pipetting combinatorial primers into
library prep plates
}
